+++ src/pymetropolis/metro_network/pedestrian_network/osm.py
@@ -83,7 +83,7 @@ class OpenStreetMapPedestrianImportStep(GeoStep, OSMStep):
     - `original_id`: OSM id of the way (note that values are generally not unique).
     - `length`: computed as geometric operation on the ways' LineString, after conversion to the
       simulation CRS.
-    - `road_type`: `highway` tag value.
+    - `edge_type`: `highway` tag value.
     - `name`: `name` tag value if any, otherwise `addr:street` tag value if any, otherwise `ref` tag
       value if any.
     """